% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_features}
\alias{add_features}
\title{Add articulatory features for data frame based on column of X-SAMPA characters}
\usage{
add_features(data, col, feature, lookup = NULL)
}
\arguments{
\item{data}{A data frame with an existing column of X-SAMPA characters}

\item{col}{A string containing the name of the column with X-SAMPA characters in \code{data}}

\item{feature}{Òne or more strings containing the features to be added}

\item{lookup}{A data frame containing a lookup table with feature values. \code{lookup} is optional;
if left blank, a lookup table will be generated using \code{update_lookup}}
}
\value{
A data frame identical to \code{data} but with columns added for the specified feature(s)
}
\description{
Add articulatory features for data frame based on column of X-SAMPA characters
}
\examples{
x <- data.frame(x=1:8, y=c('p', 'p_h', 'd', 't', 't_h', 't:', 't:_>', 't:_}'))
print(x)
y <- add_features(data=x, col='y', feature=c('place', 'lar', 'release', 'length'))
print(y)
}
