% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_reassign.R
\name{feature_reassign}
\alias{feature_reassign}
\title{Change feature values for X-SAMPA character}
\usage{
feature_reassign(sampa, feature = NA, val = NA, lookup = NULL)
}
\arguments{
\item{sampa}{String containing existing X-SAMPA characters to be modified}

\item{feature}{One or more strings specifying the features to be changed}

\item{val}{One or more strings specifying the values to be associated with the \code{feature}s}

\item{lookup}{A data frame containing a lookup table with feature values. \code{lookup} is optional;
if left blank, a lookup table will be generated using \code{update_lookup}}
}
\value{
A data frame containing a feature lookup table
}
\description{
Change feature values for X-SAMPA character
}
\examples{
lkup <- update_lookup()
lkup[lkup$segm=='t'|lkup$segm=='d',]
x <- feature_reassign(sampa='t', feature='place', val='dental')
x[x$segm=='t',]
y <- feature_reassign(sampa='d', feature='place', val='dental', lookup=x)
y[y$segm=='d',]
}
