% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{phonfeatures test data}
\format{
\subsection{\code{test_data}}{

A data frame with 3,744 rows and 6 columns.
Only the Stop column is relevant here,
but see more documentation in the link below.
}
}
\source{
\url{https://dataverse.nl/dataset.xhtml?persistentId=doi:10.34894/OSGTR8}
}
\usage{
test_data
}
\description{
A subset of data from a study on intervocalic voicing in Danish stops,
}
\keyword{datasets}
