% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_assign.R
\name{feature_assign}
\alias{feature_assign}
\title{Assign features to X-SAMPA character}
\usage{
feature_assign(new, feature = NA, val = NA, copy = NA, lookup = NULL)
}
\arguments{
\item{new}{A string containing the character to be added to the X-SAMPA
lookup table.}

\item{feature}{One or more strings specifying the features to be associated
with \code{new}.
The features for which generic features are provided are:
\itemize{
\item \code{height} vowel height
\item \code{backness} vowel backness
\item \code{roundness} vowel roundness
\item \code{place} consonant place of articulation
\item \verb{major place} major place features (coronal, dorsal, etc.)
\item \code{manner} manner of articulation
\item \code{major_manner} major manner features (obstruent, sonorant)
\item \code{lar} laryngeal features
\item \code{voice} binary voicing feature
\item \code{length}
\item \code{modifications} associated with diacritics, such as \code{velarized}, \code{palatalized}, etc.
\item \code{syllabic}
\item \code{release} modifying diacritics for (mostly) stop releases
\item \code{nasalization}
\item \code{tone}
}}

\item{val}{One or more strings specifying the values to be associated with
the \code{feature}s.}

\item{copy}{A string containing an existing character in the lookup table from
which other features are copied.}

\item{lookup}{A data frame containing a lookup table with feature values.
\code{lookup} is optional; if left blank, a lookup table will be generated using
\code{\link[=update_lookup]{update_lookup()}}.}
}
\value{
A data frame containing a feature lookup table.
}
\description{
\code{feature_assign()} is used to create a new feature lookup table with
added unknown X-SAMPA characters.
}
\details{
Note that features assigned to new characters will be inherited by
versions of that character with diacritics. For example, if the character
\code{ph} is copied from \code{p} with the value \code{aspirated} for the feature \code{lar},
then passing \code{ph_G} to \code{\link[=feature_lookup]{feature_lookup()}} will return a velarized aspirated
stop (see the final example below). If you need to override this, \code{ph_G} must
be manually assigned features.

\code{feature_reassign()} does not presently support IPA
characters. The function \code{\link[ipa:ipa]{ipa::ipa()}} is used under the hood to convert IPA
characters to X-SAMPA, and if you are using characters that are not known
to \code{\link[ipa:ipa]{ipa::ipa()}}, the functionality in this package will fail.
}
\examples{
x <- feature_assign(new='ph', feature='lar', val='aspirated', copy='p')
tail(x)
y <- feature_assign(new='th', feature='lar', val='aspirated', copy='t', lookup=x)
tail(y)
### If copy is left blank, NA's will be generated for unspecified features
z <- feature_assign(new='ph', feature='lar', val='aspirated')
tail(z)
###
feature_lookup(phon='ph_G', feature=c('lar', 'modifications'), lookup=x)
}
\seealso{
\code{\link[=update_lookup]{update_lookup()}} which generates a generic feature lookup table if
no other lookup table is provided; \code{\link[=feature_lookup]{feature_lookup()}} which is used under
the hood to look up feature values; \code{\link[=feature_reassign]{feature_reassign()}} which is used
to change the generic feature values provided by \code{\link[=update_lookup]{update_lookup()}};
\code{\link[=add_features]{add_features()}} which is used to add feature column(s) to a data frame.
}
