% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_lookup.R
\name{feature_lookup}
\alias{feature_lookup}
\title{Look up articulatory features}
\usage{
feature_lookup(
  phon,
  feature = c("height", "backness", "roundness", "place", "major_place", "manner",
    "major_manner", "lar", "voice", "length", "modifications", "syllabic", "release",
    "nasalization", "tone"),
  lookup = NULL,
  ipa = FALSE
)
}
\arguments{
\item{phon}{A string containing a phonetic character.}

\item{feature}{One or more strings specifying the features to look up;
if left blank, all pre-specified feature are returned.}

\item{lookup}{A data frame containing a lookup table with feature values.
\code{lookup} is optional; if left blank, a lookup table will be generated using
\code{\link[=update_lookup]{update_lookup()}}.}

\item{ipa}{A Boolean indicating whether the phonetic characters in \code{col} are
IPA characters. Default is \code{FALSE}.}
}
\value{
A character object or data frame containing the requested feature(s)
}
\description{
Given a phonetic character, \code{feature_lookup()} can be used to return one or
more articulatory feature values. \code{feature_lookup()} can be used to check
whether the generic feature values provided by \code{\link[=update_lookup]{update_lookup()}} are in
agreement with the user's needs, and the function is used under the hood for
the \code{\link[=add_features]{add_features()}} function.
}
\examples{
feature_lookup(phon='p_h')
feature_lookup(phon='p_h', feature='place')
x <- feature_reassign(sampa='t', feature='place', val='dental')
feature_lookup(phon='t', feature='place', lookup=x)
###
feature_lookup(phon='2', feature='roundness')
feature_lookup(phon='ø', feature='roundness', ipa=TRUE)
}
\seealso{
\code{\link[=update_lookup]{update_lookup()}} which generates a generic feature lookup table if
no other lookup table is provided; \code{\link[=feature_assign]{feature_assign()}} which is used to
assign features to unknown characters; \code{\link[=feature_reassign]{feature_reassign()}} which is used
to change the generic feature values provided by \code{\link[=update_lookup]{update_lookup()}};
\code{\link[=add_features]{add_features()}} which is used to add feature column(s) to a data frame.

\code{\link[ipa:ipa]{ipa::ipa()}} is used to convert X-SAMPA characters into IPA characters if
\code{ipa=TRUE}.
}
