% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_lookup.R
\name{feature_lookup}
\alias{feature_lookup}
\title{Look up articulatory features for X-SAMPA character}
\usage{
feature_lookup(
  sampa,
  feature = c("height", "backness", "roundness", "place", "major_place", "manner",
    "major_manner", "lar", "voice", "length", "modifications", "syllabic", "release",
    "nasalization", "tone"),
  lookup = NULL
)
}
\arguments{
\item{sampa}{A string containing an X-SAMPA character}

\item{feature}{One or more strings specifying the features to look up;
if left blank, all pre-specified feature are returned.}

\item{lookup}{A data frame containing a lookup table with feature values. \code{lookup} is optional;
if left blank, a lookup table will be generated using \code{update_lookup}}
}
\value{
A character object or data frame containing the requested feature(s)
}
\description{
Look up articulatory features for X-SAMPA character
}
\examples{
feature_lookup(sampa='p_h')
feature_lookup(sampa='p_h', feature='place')
x <- feature_reassign(sampa='t', feature='place', val='dental')
feature_lookup(sampa='t', feature='place', lookup=x)
}
