% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_reassign.R
\name{feature_reassign}
\alias{feature_reassign}
\title{Change feature values for X-SAMPA character}
\usage{
feature_reassign(sampa, feature = NA, val = NA, lookup = NULL)
}
\arguments{
\item{sampa}{One or more strings containing existing X-SAMPA characters to be modified.}

\item{feature}{One or more strings specifying the features to be changed.
The features for which generic features are provided are:
\itemize{
\item \code{height} vowel height
\item \code{backness} vowel backness
\item \code{roundness} vowel roundness
\item \code{place} consonant place of articulation
\item \code{major_place} major place features (coronal, dorsal, etc.)
\item \code{manner} manner of articulation
\item \code{major_manner} major manner features (obstruent, sonorant)
\item \code{lar} laryngeal features
\item \code{voice} binary voicing feature
\item \code{length}
\item \code{modifications} associated with diacritics, such as \code{velarized}, \code{palatalized}, etc.
\item \code{syllabic}
\item \code{release} modifying diacritics for (mostly) stop releases
\item \code{nasalization}
\item \code{tone}
}}

\item{val}{One or more strings specifying the values to be associated with
the \code{feature}s.}

\item{lookup}{A data frame containing a lookup table with feature values.
\code{lookup} is optional; if left blank, a lookup table will be generated using
\code{\link[=update_lookup]{update_lookup()}}.}
}
\value{
A data frame containing a feature lookup table.
}
\description{
\code{feature_reassign()} is used to create a new feature lookup table for X-SAMPA
characters if the user wishes to assign different (non-standard) feature
values to existing characters.
}
\details{
Note that reassigned features will be inherited by any characters
with diacritics. For example, if the character \code{t} is assigned the place
feature \code{dental}, then \code{t_h} will also be assigned the place feature
\code{dental}. If you need to override this, \code{t_h} can be assigned manually using
\code{\link[=feature_assign]{feature_assign()}}.

\code{sampa} can take either a string or multiple strings, if the same feature
needs to be changed for multiple characters. See example below.

\code{feature_reassign()} does not explicitly support IPA
characters. However, if \code{feature_reassign()} is used on an X-SAMPA string
corresponding to the IPA string you wish to modify, \code{\link[=add_features]{add_features()}} with
\code{ipa=TRUE} should still give the desired result (see the last example).
}
\examples{
lkup <- update_lookup()
lkup[lkup$segm=='t'|lkup$segm=='d',]
x <- feature_reassign(sampa='t', feature=c('lar','place'),
                      val=c('aspirated','dental'))
x[x$segm=='t',]
y <- feature_reassign(sampa='d', feature='place', val='dental', lookup=x)
y[y$segm=='d',]
### Alternatively,
x <- feature_reassign(sampa=c('d', 't'), feature='place', val='dental')
###
feature_lookup(phon='d', feature='place', lookup=x, ipa=TRUE)
}
\seealso{
\code{\link[=update_lookup]{update_lookup()}} which generates a generic feature lookup table if
no other lookup table is provided; \code{\link[=feature_assign]{feature_assign()}} which is used to
assign features to unknown characters; \code{\link[=feature_lookup]{feature_lookup()}} which is used under
the hood to look up feature values. \code{\link[=add_features]{add_features()}} which is used to add
feature column(s) to a data frame.
}
