% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_reassign.R
\name{feature_reassign}
\alias{feature_reassign}
\title{Change feature values for X-SAMPA character}
\usage{
feature_reassign(sampa, feature = NA, val = NA, lookup = NULL)
}
\arguments{
\item{sampa}{String containing existing X-SAMPA characters to be modified.}

\item{feature}{One or more strings specifying the features to be changed.}

\item{val}{One or more strings specifying the values to be associated with
the \code{feature}s.}

\item{lookup}{A data frame containing a lookup table with feature values.
\code{lookup} is optional; if left blank, a lookup table will be generated using
\code{\link[=update_lookup]{update_lookup()}}.}
}
\value{
A data frame containing a feature lookup table.
}
\description{
\code{feature_reassign()} is used to create a new feature lookup table for X-SAMPA
characters if the user wishes to assign different (non-standard) feature
values to existing characters.
}
\details{
Note that reassigned features will be inherited by any characters
with diacritics. For example, if the character \code{t} is assigned the place
feature \code{dental}, then \code{t_h} will also be assigned the place feature
\code{dental}. If you need to override this, \code{t_h} can be assigned manually using
\code{\link[=feature_assign]{feature_assign()}}.

\code{feature_reassign()} does not explicitly support IPA
characters. However, if \code{feature_reassign()} is used on an X-SAMPA string
corresponding to the IPA string you wish to modify, \code{\link[=add_features]{add_features()}} with
\code{ipa=TRUE} should still give the desired result (see the last example).
}
\examples{
lkup <- update_lookup()
lkup[lkup$segm=='t'|lkup$segm=='d',]
x <- feature_reassign(sampa='t', feature='place', val='dental')
x[x$segm=='t',]
y <- feature_reassign(sampa='d', feature='place', val='dental', lookup=x)
y[y$segm=='d',]
###
feature_lookup(phon='d', feature='place', lookup=x, ipa=TRUE)
}
\seealso{
\code{\link[=update_lookup]{update_lookup()}} which generates a generic feature lookup table if
no other lookup table is provided; \code{\link[=feature_assign]{feature_assign()}} which is used to
assign features to unknown characters; \code{\link[=feature_lookup]{feature_lookup()}} which is used under
the hood to look up feature values. \code{\link[=add_features]{add_features()}} which is used to add
feature column(s) to a data frame.
}
