% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lookup.R
\name{update_lookup}
\alias{update_lookup}
\title{Update feature lookup file with redundant features}
\usage{
update_lookup()
}
\value{
A data frame
}
\description{
Update feature lookup file with redundant features
}
\examples{
x <- update_lookup()
}
