% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lookup.R
\name{update_lookup}
\alias{update_lookup}
\title{Create feature lookup table with generic features}
\usage{
update_lookup()
}
\value{
A data frame containing a feature lookup table
}
\description{
Create feature lookup table with generic features
}
\examples{
x <- update_lookup()
}
